+++ kotlin-stdlib/common/test/kotlin/com/javiersc/kotlin/stdlib/StringsTest.kt
@@ -1,23 +1,21 @@
 package com.javiersc.kotlin.stdlib
 
-import io.kotest.matchers.booleans.shouldBeFalse
-import io.kotest.matchers.booleans.shouldBeTrue
-import io.kotest.matchers.shouldBe
-import io.kotest.matchers.string.shouldBeEmpty
 import kotlin.test.Test
+import kotlin.test.assertFalse
+import kotlin.test.assertTrue
 
 class StringsTest {
 
     @Test
     fun string_remove() {
-        "Hello, World".remove("Hello, ").shouldBe("World")
-        "Hello, World".remove("bla").shouldBe("Hello, World")
-        "Hello, World".remove("llo", "rld").shouldBe("He, Wo")
+        assertTrue { "Hello, World".remove("Hello, ") == "World" }
+        assertTrue { "Hello, World".remove("bla") == "Hello, World" }
+        assertTrue { "Hello, World".remove("llo", "rld") == "He, Wo" }
     }
 
     @Test
     fun string_replace() {
-        "Hello, World".replace("ello" to "ELLO", "orld" to "ORLD").shouldBe("HELLO, WORLD")
+        assertTrue { "Hello, World".replace("ello" to "ELLO", "orld" to "ORLD") == "HELLO, WORLD" }
     }
 
     @Test
@@ -27,36 +25,36 @@ class StringsTest {
         val empty = ""
         val notBlank = "Hello, World"
 
-        nullable.isNotNullNorBlank().shouldBeFalse()
-        nullable.isNotNullNorEmpty().shouldBeFalse()
-        blank.isNotNullNorBlank().shouldBeFalse()
-        blank.isNotNullNorEmpty().shouldBeTrue()
-        empty.isNotNullNorBlank().shouldBeFalse()
-        empty.isNotNullNorEmpty().shouldBeFalse()
-        notBlank.isNotNullNorBlank().shouldBeTrue()
-        notBlank.isNotNullNorEmpty().shouldBeTrue()
+        assertFalse { nullable.isNotNullNorBlank() }
+        assertFalse { nullable.isNotNullNorEmpty() }
+        assertFalse { blank.isNotNullNorBlank() }
+        assertTrue { blank.isNotNullNorEmpty() }
+        assertFalse { empty.isNotNullNorBlank() }
+        assertFalse { empty.isNotNullNorEmpty() }
+        assertTrue { notBlank.isNotNullNorBlank() }
+        assertTrue { notBlank.isNotNullNorEmpty() }
     }
 
     @Test
     fun empty_string() {
-        String.Empty.shouldBeEmpty()
+        assertTrue { String.Empty == "" }
     }
 
     @Test
     fun remove_duplicate_empty_lines() {
-        "a\nb\n\n\nc\n".removeDuplicateEmptyLines().shouldBe("a\nb\n\nc\n")
-        "a\n\nb\n\n\nc\n".removeDuplicateEmptyLines().shouldBe("a\n\nb\n\nc\n")
-        "a\n\nb\n\n\nc\n\n".removeDuplicateEmptyLines().shouldBe("a\n\nb\n\nc\n")
-        "a\n\n\n\n\nb\n\n".removeDuplicateEmptyLines().shouldBe("a\n\nb\n")
+        assertTrue { "a\nb\n\n\nc\n".removeDuplicateEmptyLines() == "a\nb\n\nc\n" }
+        assertTrue { "a\n\nb\n\n\nc\n".removeDuplicateEmptyLines() == "a\n\nb\n\nc\n" }
+        assertTrue { "a\n\nb\n\n\nc\n\n".removeDuplicateEmptyLines() == "a\n\nb\n\nc\n" }
+        assertTrue { "a\n\n\n\n\nb\n\n".removeDuplicateEmptyLines() == "a\n\nb\n" }
     }
 
     @Test
     fun end_with_new_line() {
-        "a".endWithNewLine().shouldBe("a\n")
-        "a\n".endWithNewLine().shouldBe("a\n")
-        "".endWithNewLine().shouldBe("")
-        "a\nb".endWithNewLine().shouldBe("a\nb\n")
-        "".endWithNewLine().shouldBe("")
-        "\n".endWithNewLine().shouldBe("\n")
+        assertTrue { "a".endWithNewLine() == "a\n" }
+        assertTrue { "a\n".endWithNewLine() == "a\n" }
+        assertTrue { "".endWithNewLine() == "" }
+        assertTrue { "a\nb".endWithNewLine() == "a\nb\n" }
+        assertTrue { "".endWithNewLine() == "" }
+        assertTrue { "\n".endWithNewLine() == "\n" }
     }
 }