+++ kotlin-test-junit5/common/main/kotlin/com/javiersc/kotlin/test/IgnoreJS.kt
@@ -1,0 +1,3 @@
+package com.javiersc.kotlin.test
+
+@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION) public expect annotation class IgnoreJS