+++ kotlin-test-junit5/ios/main/kotlin/com/javiersc/kotlin/test/IgnoreTVOSX64.kt
@@ -1,0 +1,4 @@
+package com.javiersc.kotlin.test
+
+@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
+public actual annotation class IgnoreTVOSX64