+++ kotlin-test-junit5/jvm/main/kotlin/com/javiersc/kotlin/test/IgnoreJVM.kt
@@ -1,0 +1,5 @@
+package com.javiersc.kotlin.test
+
+import org.junit.jupiter.api.Disabled
+
+public actual typealias IgnoreJVM = Disabled