+++ kotlin-test-junit5/mingw/main/kotlin/com/javiersc/kotlin/test/IgnoreMINGW.kt
@@ -1,0 +1,5 @@
+package com.javiersc.kotlin.test
+
+import kotlin.test.Ignore
+
+public actual typealias IgnoreMINGW = Ignore