+++ kotlin-test-junit5/tvosArm64/main/kotlin/com/javiersc/kotlin/test/IgnoreTVOSARM64.kt
@@ -1,0 +1,5 @@
+package com.javiersc.kotlin.test
+
+import kotlin.test.Ignore
+
+public actual typealias IgnoreTVOSARM64 = Ignore