+++ kotlin-test-junit5/wasm/main/kotlin/com/javiersc/kotlin/test/IgnoreIOS.kt
@@ -1,0 +1,3 @@
+package com.javiersc.kotlin.test
+
+@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION) public actual annotation class IgnoreIOS